fix: fixed all warnings

+++ src/result.hpp
@@ -195,7 +195,7 @@ public:
     }
 
 private:
-    int _stars = 0;
+    size_t _stars = 0;
     void _get_is_3stars()
     {
         auto& self = *this;
@@ -508,14 +508,14 @@ private:
         auto& self = *this;
         self._relate(diff_rect.tl());
 
-        int diff_count = 0;
+        size_t diff_count = 0;
 
         if (!(height < width / 4))
         {
             img_bin = _img;
             cv::cvtColor(img_bin, img_bin, cv::COLOR_BGR2GRAY);
             cv::threshold(img_bin, img_bin, 200, 255, cv::THRESH_BINARY);
-            int inspction_line = 0.8 * cv::boundingRect(img_bin).y;
+            int inspction_line = static_cast<int>(0.8 * cv::boundingRect(img_bin).y);
 
             img_bin = _img;
             cv::cvtColor(img_bin, img_bin, cv::COLOR_BGR2GRAY);
@@ -602,7 +602,7 @@ private:
                 {
                     continue;
                 }
-                int dist = abs(kh - hsv[H]);
+                int dist = static_cast<int>(abs(kh - hsv[H]));
                 _candidates.emplace_back(vtype, dist);
             }
             std::sort(_candidates.begin(), _candidates.end(),
@@ -712,7 +712,7 @@ private:
         while (droptextimg.rows > 0)
         {
             cv::Mat topline = droptextimg.row(0);
-            int meanval = cv::mean(topline)[0];
+            int meanval = static_cast<int>(cv::mean(topline)[0]);
             if (meanval < 127)
             {
                 break;
@@ -780,7 +780,7 @@ public:
     }
 
 private:
-    int _items_count = round(width / (height * W_H_PROP));
+    int _items_count = static_cast<int>(round(width / (height * W_H_PROP)));
     Widget_DroptypeLine _line {this};
     Widget_DroptypeText _text {this};
     void _get_candidates()
@@ -859,12 +859,12 @@ public:
         rpt["dropTypes"] = dict::array();
         rpt["drops"] = dict::array();
 
-        int droptypes_count = _droptype_list.size(); // will move in "for" in C++20
+        size_t droptypes_count = _droptype_list.size(); // will move in "for" in C++20
         for (int i = 0; i < droptypes_count; i++)
         {
             rpt["dropTypes"].push_back(_droptype_list[i].report(debug));
         }
-        int drops_count = _drop_list.size(); // will move in "for" in C++20
+        size_t drops_count = _drop_list.size(); // will move in "for" in C++20
         for (int i = 0; i < drops_count; i++)
         {
             rpt["drops"].push_back(
@@ -882,7 +882,7 @@ private:
     auto _get_separate()
     {
         cv::Mat img_bin = _img;
-        int offset = round(height * 0.75);
+        int offset = static_cast<int>(round(height * 0.75));
         img_bin.adjustROI(-offset, 0, 0, 0);
         cv::cvtColor(img_bin, img_bin, cv::COLOR_BGR2GRAY);
         cv::adaptiveThreshold(img_bin, img_bin, 255, cv::ADAPTIVE_THRESH_MEAN_C,
@@ -911,7 +911,7 @@ private:
         }
         int baseline_h = row + offset;
         auto sp = separate(img_bin(cv::Rect(0, row, width, 1)), DirectionFlags::LEFT);
-        int item_diameter = height / DROP_AREA_HEIGHT_PROP * ITEM_DIAMETER_PROP;
+        int item_diameter = static_cast<int>(height / DROP_AREA_HEIGHT_PROP * ITEM_DIAMETER_PROP);
         for (auto it = sp.cbegin(); it != sp.cend();)
         {
             const auto& range = *it;
@@ -991,7 +991,7 @@ private:
         {
             return;
         }
-        int item_diameter = height / DROP_AREA_HEIGHT_PROP * ITEM_DIAMETER_PROP;
+        int item_diameter = static_cast<int>(height / DROP_AREA_HEIGHT_PROP * ITEM_DIAMETER_PROP);
         ItemTemplates templs {stage};
         for (const auto& droptype : _droptype_list)
         {
@@ -1206,8 +1206,8 @@ private:
         }
         const auto& bv = _baseline_v;
         auto drop_area_img = _img(
-            cv::Range(bv.y + bv.height * DROP_AREA_Y_PROP, bv.y + bv.height),
-            cv::Range(bv.x + bv.height * DROP_AREA_X_PROP, width));
+            cv::Range(static_cast<int>(bv.y + bv.height * DROP_AREA_Y_PROP), bv.y + bv.height),
+            cv::Range(static_cast<int>(bv.x + bv.height * DROP_AREA_X_PROP), width));
         _drop_area.set_img(drop_area_img);
         _drop_area.analyze(_stage.stage_code(), "NORMAL");
     }
@@ -1275,7 +1275,11 @@ private:
         {
             return;
         }
-        cv::Mat img_bin = _img(cv::Rect(0, 0.2 * height, 0.2 * width, 0.4 * height));
+        cv::Mat img_bin = _img(cv::Rect(
+            0,
+            static_cast<int>(0.2 * height),
+            static_cast<int>(0.2 * width),
+            static_cast<int>(0.4 * height)));
         cv::cvtColor(img_bin, img_bin, cv::COLOR_BGR2GRAY);
         cv::threshold(img_bin, img_bin, 120, 255, cv::THRESH_BINARY);
 
@@ -1291,7 +1295,7 @@ private:
                 abs(img_temp.cols - last_height) <= 1 &&
                 abs(first_height - last_height) <= 1)
             {
-                baseline_v_rect = cv::Rect(range.start, sp2.front().start + 0.2 * height,
+                baseline_v_rect = cv::Rect(range.start, sp2.front().start + static_cast<int>(0.2 * height),
                                            img_temp.cols, sp2.back().end - sp2.front().start);
                 break;
             }
@@ -1318,7 +1322,11 @@ private:
                 : x(x_), y(y_), width(width_), height(height_), area(area_) {}
         };
 
-        cv::Mat img_bin = _img(cv::Rect(0, 0.2 * height, 0.2 * width, 0.4 * height));
+        cv::Mat img_bin = _img(cv::Rect(
+            0,
+            static_cast<int>(0.2 * height),
+            static_cast<int>(0.2 * width),
+            static_cast<int>(0.4 * height)));
         cv::cvtColor(img_bin, img_bin, cv::COLOR_BGR2GRAY);
         cv::threshold(img_bin, img_bin, 120, 255, cv::THRESH_BINARY);
         cv::Mat _;
@@ -1360,7 +1368,7 @@ private:
             if (ccomps[0].x == ccomps[1].x)
             {
                 baseline_v_rect = cv::Rect(ccomps[0].x,
-                                           ccomps[0].y + 0.2 * height,
+                                           ccomps[0].y + static_cast<int>(0.2 * height),
                                            ccomps[0].width,
                                            ccomps[1].y - ccomps[0].y + ccomps[1].height);
             }
@@ -1372,7 +1380,7 @@ private:
             {
                 if (ccomps[i].x == ccomps[i + 1].x)
                     baseline_v_rect = cv::Rect(ccomps[i].x,
-                                               ccomps[i].y + 0.2 * height,
+                                               ccomps[i].y + static_cast<int>(0.2 * height),
                                                ccomps[i].width,
                                                ccomps[i + 1].y - ccomps[i].y + ccomps[i + 1].height);
             }
@@ -1399,7 +1407,7 @@ private:
         const auto& bv = _baseline_v;
         int left_margin = bv.x + bv.width;
         auto result_img = _img(cv::Range(bv.y, bv.y + bv.height / 2),
-                               cv::Range(left_margin, left_margin + 1.5 * bv.height));
+                               cv::Range(left_margin, static_cast<int>(left_margin + 1.5 * bv.height)));
         cv::Mat img_bin;
         cv::cvtColor(result_img, img_bin, cv::COLOR_BGR2GRAY);
         cv::threshold(img_bin, img_bin, 200, 255, cv::THRESH_BINARY);
@@ -1417,7 +1425,7 @@ private:
         const auto& bv = _baseline_v;
         int left_margin = bv.x + bv.width;
         auto star_img = _img(cv::Range(bv.y + bv.height / 2, bv.y + bv.height),
-                             cv::Range(left_margin, left_margin + 1.2 * bv.height));
+                             cv::Range(left_margin, static_cast<int>(left_margin + 1.2 * bv.height)));
         cv::Mat img_bin;
         cv::cvtColor(star_img, img_bin, cv::COLOR_BGR2GRAY);
         cv::threshold(img_bin, img_bin, 127, 255, cv::THRESH_BINARY);
@@ -1435,8 +1443,8 @@ private:
         const auto& bv = _baseline_v;
         int left_margin = bv.x + bv.width;
         auto stage_img = _img(cv::Range(0, bv.y),
-                              cv::Range(left_margin + 0.43 * bv.height,
-                                        left_margin + 1.5 * bv.height));
+                              cv::Range(static_cast<int>(left_margin + 0.43 * bv.height),
+                                        static_cast<int>(left_margin + 1.5 * bv.height)));
         cv::Mat img_bin;
         cv::cvtColor(stage_img, img_bin, cv::COLOR_BGR2GRAY);
         cv::threshold(img_bin, img_bin, 200, 255, cv::THRESH_BINARY);
@@ -1454,7 +1462,7 @@ private:
         const auto& bv = _baseline_v;
         int left_margin = bv.x + bv.width;
         auto diff_img = _img(cv::Range(0, bv.y),
-                             cv::Range(left_margin, left_margin + 0.43 * bv.height));
+                             cv::Range(left_margin, static_cast<int>(left_margin + 0.43 * bv.height)));
         cv::Mat img_bin;
         cv::cvtColor(diff_img, img_bin, cv::COLOR_BGR2GRAY);
         cv::threshold(img_bin, img_bin, 25, 255, cv::THRESH_BINARY);
@@ -1471,19 +1479,18 @@ private:
         }
         const auto& bv = _baseline_v;
         cv::Mat img_bin = _img(cv::Range(bv.y + bv.height, height),
-                               cv::Range(bv.x + bv.width, 0.2 * width));
+                               cv::Range(bv.x + bv.width, static_cast<int>(0.2 * width)));
         cv::cvtColor(img_bin, img_bin, cv::COLOR_BGR2GRAY);
         cv::threshold(img_bin, img_bin, 127, 255, cv::THRESH_BINARY);
         auto sp = separate(img_bin, DirectionFlags::TOP, 2);
         int top_margin = bv.y + bv.height + sp[1].start;
         auto drop_area_img = _img(
-            cv::Range(top_margin + bv.height * DROP_AREA_Y_PROP, top_margin + bv.height),
+            cv::Range(top_margin + static_cast<int>(bv.height * DROP_AREA_Y_PROP), top_margin + bv.height),
             cv::Range(bv.x + bv.width, width));
         _drop_area.set_img(drop_area_img);
         _drop_area.analyze(_stage.stage_code(), _difficutly.difficulty());
     }
 };
-
 } // namespace penguin
 
 #endif // PENGUIN_RESULT_HPP_
